requirements.txt
yt_dlp 
git+https://github.com/Hrishi2861/pyrofork-2.2.11-peer-fix.git 
aria2 
TgCrypto
static_ffmpeg
speedtest-cli
flask
bs4
python-dotenv
psutil